# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "tssk-management ",
        "description": "Task Management Swagger api documentation",
        "contact": {
            "email": ""
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login user",
                "description": "The process of gaining access to a website by providing valid credentials.",
                "operationId": "login",
                "requestBody": {
                    "description": "Pass user data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "example": "admin@gmail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Mahmaud Arshad Nahid"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "admin@gmail.com"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2024-02-05T07:52:41.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2024-02-05T07:52:41.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "token_details": {
                                            "properties": {
                                                "access_token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vdGFzay1tYW5hZ2VtZW50LnRlc3QvYXBpL2xvZ2luIiwiaWF0IjoxNzEzOTUzNTk3LCJleHAiOjE3MTM5NTcxOTcsIm5iZiI6MTcxMzk1MzU5NywianRpIjoiNzVkY2FYYlZYUmFVYzVYSyIsInN1YiI6IjExIiwicHJ2IjoiMjNiZDVjODk0OWY2MDBhZGIzOWU3MDFjNDAwODcyZGI3YTU5NzZmNyJ9.vD4a7RL-ElDK8u1PN1o1anC_N77oJXVKhxd0JvlC3zI"
                                                },
                                                "token_type": {
                                                    "type": "string",
                                                    "example": "bearer"
                                                },
                                                "expires_in": {
                                                    "type": "string",
                                                    "example": "3600"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Login successful!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "[]"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Credentials do not match"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error Has Occurred",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "data": {
                                            "properties": {
                                                "email": {
                                                    "type": "string",
                                                    "example": "The email field is required."
                                                },
                                                "email.exists": {
                                                    "type": "string",
                                                    "example": "Invalid Email."
                                                },
                                                "password": {
                                                    "type": "string",
                                                    "example": "The password field is required."
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation Error Has Occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Data not found!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": "[]"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data not found!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "description": "Logout",
                "operationId": "e6a758d5d5373f9421674724f39f3e57",
                "requestBody": {
                    "description": "Logout from the device",
                    "required": false
                },
                "responses": {
                    "200": {
                        "description": "success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": null
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully Logout"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": null
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An Error Has Occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Login with email and password to get the authentication token Like Bearer 10|4yB0GEn8fX9JHtXGwT0hOckxIqDkNiiaXmzwOTw36c325624",
                "name": "Token based Based",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}